move back::link::write into a separate file
